test: reorganize by concern
Reorganize tests by concern, use table-driven tests,
increase coverage.
