+++ callsite_test.go
@@ -1,7 +1,6 @@
 package resurgo_test
 
 import (
-	"bytes"
 	"debug/elf"
 	"os"
 	"os/exec"
@@ -29,77 +28,70 @@ func TestDetectCallSitesAMD64_Call(t *testing.T) {
 		wantConf   resurgo.Confidence
 		wantSource uint64
 		wantTarget uint64
-	}{
-		{
-			name: "pc-relative-call",
-			// call $+0x10 (rel32 = 0x0000000B, instruction length = 5)
-			// Target = 0 + 5 + 0x0B = 0x10
-			code:       []byte{0xE8, 0x0B, 0x00, 0x00, 0x00},
-			baseAddr:   0,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantMode:   resurgo.AddressingModePCRelative,
-			wantConf:   resurgo.ConfidenceHigh,
-			wantSource: 0,
-			wantTarget: 0x10,
-		},
-		{
-			name: "pc-relative-call-negative-offset",
-			// call $-0x20 (rel32 = 0xFFFFFFE0, two's complement -32)
-			// At address 0x100, target = 0x100 + 5 + (-32) = 0xE5
-			code:       []byte{0xE8, 0xE0, 0xFF, 0xFF, 0xFF},
-			baseAddr:   0x100,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantMode:   resurgo.AddressingModePCRelative,
-			wantConf:   resurgo.ConfidenceHigh,
-			wantSource: 0x100,
-			wantTarget: 0xE5,
-		},
-		{
-			name: "register-indirect-call",
-			// call rax = FF D0
-			code:       []byte{0xFF, 0xD0},
-			baseAddr:   0x200,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantMode:   resurgo.AddressingModeRegisterIndirect,
-			wantConf:   resurgo.ConfidenceNone,
-			wantSource: 0x200,
-			wantTarget: 0,
-		},
-		{
-			name: "rip-relative-call",
-			// call [rip+0x1234] = FF 15 34 12 00 00 (6 bytes)
-			// At address 0x1000, target = 0x1000 + 6 + 0x1234 = 0x223A
-			code:       []byte{0xFF, 0x15, 0x34, 0x12, 0x00, 0x00},
-			baseAddr:   0x1000,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantMode:   resurgo.AddressingModePCRelative,
-			wantConf:   resurgo.ConfidenceMedium,
-			wantSource: 0x1000,
-			wantTarget: 0x223A,
-		},
-		{
-			name: "memory-call-with-base-register",
-			// call [rbx+0x10] = FF 53 10
-			code:       []byte{0xFF, 0x53, 0x10},
-			baseAddr:   0x300,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantMode:   resurgo.AddressingModeRegisterIndirect,
-			wantConf:   resurgo.ConfidenceNone,
-			wantSource: 0x300,
-			wantTarget: 0,
-		},
-		{
-			name:      "no-call-instructions",
-			code:      []byte{0x90, 0x90, 0x90}, // nop, nop, nop
-			baseAddr:  0,
-			wantCount: 0,
-		},
-	}
+	}{{
+		name: "pc-relative-call",
+		// call $+0x10 (rel32 = 0x0000000B, instruction length = 5)
+		// Target = 0 + 5 + 0x0B = 0x10
+		code:       []byte{0xE8, 0x0B, 0x00, 0x00, 0x00},
+		baseAddr:   0,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantMode:   resurgo.AddressingModePCRelative,
+		wantConf:   resurgo.ConfidenceHigh,
+		wantSource: 0,
+		wantTarget: 0x10,
+	}, {
+		name: "pc-relative-call-negative-offset",
+		// call $-0x20 (rel32 = 0xFFFFFFE0, two's complement -32)
+		// At address 0x100, target = 0x100 + 5 + (-32) = 0xE5
+		code:       []byte{0xE8, 0xE0, 0xFF, 0xFF, 0xFF},
+		baseAddr:   0x100,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantMode:   resurgo.AddressingModePCRelative,
+		wantConf:   resurgo.ConfidenceHigh,
+		wantSource: 0x100,
+		wantTarget: 0xE5,
+	}, {
+		name: "register-indirect-call",
+		// call rax = FF D0
+		code:       []byte{0xFF, 0xD0},
+		baseAddr:   0x200,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantMode:   resurgo.AddressingModeRegisterIndirect,
+		wantConf:   resurgo.ConfidenceNone,
+		wantSource: 0x200,
+		wantTarget: 0,
+	}, {
+		name: "rip-relative-call",
+		// call [rip+0x1234] = FF 15 34 12 00 00 (6 bytes)
+		// At address 0x1000, target = 0x1000 + 6 + 0x1234 = 0x223A
+		code:       []byte{0xFF, 0x15, 0x34, 0x12, 0x00, 0x00},
+		baseAddr:   0x1000,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantMode:   resurgo.AddressingModePCRelative,
+		wantConf:   resurgo.ConfidenceMedium,
+		wantSource: 0x1000,
+		wantTarget: 0x223A,
+	}, {
+		name: "memory-call-with-base-register",
+		// call [rbx+0x10] = FF 53 10
+		code:       []byte{0xFF, 0x53, 0x10},
+		baseAddr:   0x300,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantMode:   resurgo.AddressingModeRegisterIndirect,
+		wantConf:   resurgo.ConfidenceNone,
+		wantSource: 0x300,
+		wantTarget: 0,
+	}, {
+		name:      "no-call-instructions",
+		code:      []byte{0x90, 0x90, 0x90}, // nop, nop, nop
+		baseAddr:  0,
+		wantCount: 0,
+	}}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -150,43 +142,39 @@ func TestDetectCallSitesAMD64_Jump(t *testing.T) {
 		wantMode   resurgo.AddressingMode
 		wantConf   resurgo.Confidence
 		wantTarget uint64
-	}{
-		{
-			name: "unconditional-jmp-rel32",
-			// jmp $+0x20 (rel32 = 0x0000001B, instruction length = 5)
-			// Target = 0 + 5 + 0x1B = 0x20
-			code:       []byte{0xE9, 0x1B, 0x00, 0x00, 0x00},
-			baseAddr:   0,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteJump,
-			wantMode:   resurgo.AddressingModePCRelative,
-			wantConf:   resurgo.ConfidenceMedium, // Unconditional = medium
-			wantTarget: 0x20,
-		},
-		{
-			name: "unconditional-jmp-rel8",
-			// jmp $+0x10 (rel8 = 0x0E, instruction length = 2)
-			// Target = 0 + 2 + 0x0E = 0x10
-			code:       []byte{0xEB, 0x0E},
-			baseAddr:   0,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteJump,
-			wantMode:   resurgo.AddressingModePCRelative,
-			wantConf:   resurgo.ConfidenceMedium,
-			wantTarget: 0x10,
-		},
-		{
-			name: "register-indirect-jmp",
-			// jmp rax = FF E0
-			code:       []byte{0xFF, 0xE0},
-			baseAddr:   0x400,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteJump,
-			wantMode:   resurgo.AddressingModeRegisterIndirect,
-			wantConf:   resurgo.ConfidenceNone,
-			wantTarget: 0,
-		},
-	}
+	}{{
+		name: "unconditional-jmp-rel32",
+		// jmp $+0x20 (rel32 = 0x0000001B, instruction length = 5)
+		// Target = 0 + 5 + 0x1B = 0x20
+		code:       []byte{0xE9, 0x1B, 0x00, 0x00, 0x00},
+		baseAddr:   0,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteJump,
+		wantMode:   resurgo.AddressingModePCRelative,
+		wantConf:   resurgo.ConfidenceMedium, // Unconditional = medium
+		wantTarget: 0x20,
+	}, {
+		name: "unconditional-jmp-rel8",
+		// jmp $+0x10 (rel8 = 0x0E, instruction length = 2)
+		// Target = 0 + 2 + 0x0E = 0x10
+		code:       []byte{0xEB, 0x0E},
+		baseAddr:   0,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteJump,
+		wantMode:   resurgo.AddressingModePCRelative,
+		wantConf:   resurgo.ConfidenceMedium,
+		wantTarget: 0x10,
+	}, {
+		name: "register-indirect-jmp",
+		// jmp rax = FF E0
+		code:       []byte{0xFF, 0xE0},
+		baseAddr:   0x400,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteJump,
+		wantMode:   resurgo.AddressingModeRegisterIndirect,
+		wantConf:   resurgo.ConfidenceNone,
+		wantTarget: 0,
+	}}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -232,31 +220,28 @@ func TestDetectCallSitesARM64_BL(t *testing.T) {
 		wantType   resurgo.CallSiteType
 		wantConf   resurgo.Confidence
 		wantTarget uint64
-	}{
-		{
-			name: "bl-forward",
-			// BL +0x1000 (offset = 0x1000, encoded as 0x1000/4 = 0x400)
-			// Instruction: 0x94000400
-			code:       arm64Insn(0x94000400),
-			baseAddr:   0x1000,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantConf:   resurgo.ConfidenceHigh,
-			wantTarget: 0x2000, // 0x1000 + 0x1000
-		},
-		{
-			name: "bl-backward",
-			// BL -0x100 (offset = -0x100, encoded as (-0x100/4) & 0x3FFFFFF)
-			// Two's complement: 0x3FFFFFC0
-			// Instruction: 0x97FFFFC0
-			code:       arm64Insn(0x97FFFFC0),
-			baseAddr:   0x2000,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteCall,
-			wantConf:   resurgo.ConfidenceHigh,
-			wantTarget: 0x1F00, // 0x2000 - 0x100
-		},
-	}
+	}{{
+		name: "bl-forward",
+		// BL +0x1000 (offset = 0x1000, encoded as 0x1000/4 = 0x400)
+		// Instruction: 0x94000400
+		code:       arm64Insn(0x94000400),
+		baseAddr:   0x1000,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantConf:   resurgo.ConfidenceHigh,
+		wantTarget: 0x2000, // 0x1000 + 0x1000
+	}, {
+		name: "bl-backward",
+		// BL -0x100 (offset = -0x100, encoded as (-0x100/4) & 0x3FFFFFF)
+		// Two's complement: 0x3FFFFFC0
+		// Instruction: 0x97FFFFC0
+		code:       arm64Insn(0x97FFFFC0),
+		baseAddr:   0x2000,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteCall,
+		wantConf:   resurgo.ConfidenceHigh,
+		wantTarget: 0x1F00, // 0x2000 - 0x100
+	}}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -299,19 +284,17 @@ func TestDetectCallSitesARM64_B(t *testing.T) {
 		wantType   resurgo.CallSiteType
 		wantConf   resurgo.Confidence
 		wantTarget uint64
-	}{
-		{
-			name: "b-unconditional",
-			// B +0x100 (offset = 0x100, encoded as 0x100/4 = 0x40)
-			// Instruction: 0x14000040
-			code:       arm64Insn(0x14000040),
-			baseAddr:   0x1000,
-			wantCount:  1,
-			wantType:   resurgo.CallSiteJump,
-			wantConf:   resurgo.ConfidenceMedium,
-			wantTarget: 0x1100, // 0x1000 + 0x100
-		},
-	}
+	}{{
+		name: "b-unconditional",
+		// B +0x100 (offset = 0x100, encoded as 0x100/4 = 0x40)
+		// Instruction: 0x14000040
+		code:       arm64Insn(0x14000040),
+		baseAddr:   0x1000,
+		wantCount:  1,
+		wantType:   resurgo.CallSiteJump,
+		wantConf:   resurgo.ConfidenceMedium,
+		wantTarget: 0x1100, // 0x1000 + 0x100
+	}}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -383,24 +366,28 @@ func TestDetectCallSites_EmptyInput(t *testing.T) {
 	tests := []struct {
 		name string
 		code []byte
-	}{
-		{name: "nil", code: nil},
-		{name: "empty", code: []byte{}},
-	}
+		arch resurgo.Arch
+	}{{
+		name: "nil/amd64",
+		code: nil,
+		arch: resurgo.ArchAMD64,
+	}, {
+		name: "nil/arm64",
+		code: nil,
+		arch: resurgo.ArchARM64,
+	}, {
+		name: "empty/amd64",
+		code: []byte{},
+		arch: resurgo.ArchAMD64,
+	}, {
+		name: "empty/arm64",
+		code: []byte{},
+		arch: resurgo.ArchARM64,
+	}}
 
 	for _, tt := range tests {
-		t.Run(tt.name+"/amd64", func(t *testing.T) {
-			edges, err := resurgo.DetectCallSites(tt.code, 0, resurgo.ArchAMD64)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			if len(edges) != 0 {
-				t.Errorf("expected 0 edges, got %d", len(edges))
-			}
-		})
-
-		t.Run(tt.name+"/arm64", func(t *testing.T) {
-			edges, err := resurgo.DetectCallSites(tt.code, 0, resurgo.ArchARM64)
+		t.Run(tt.name, func(t *testing.T) {
+			edges, err := resurgo.DetectCallSites(tt.code, 0, tt.arch)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
 			}
@@ -437,103 +424,6 @@ func TestDetectCallSites_JumpTarget(t *testing.T) {
 	}
 	if found == nil {
 		t.Fatal("expected jump edge to 0x10, got none")
-	}
-}
-
-func TestDetectFunctionsFromELF(t *testing.T) {
-	binPath := filepath.Join(t.TempDir(), "demo-app")
-	args := []string{"build", "-o", binPath, "testdata/demo-app.go"}
-
-	cmd := exec.Command("go", args...)
-	cmd.Env = append(os.Environ(), "CGO_ENABLED=0", "GOARCH=amd64")
-	if out, err := cmd.CombinedOutput(); err != nil {
-		t.Fatalf("failed to compile demo-app: %v\n%s", err, out)
-	}
-
-	f, err := elf.Open(binPath)
-	if err != nil {
-		t.Fatalf("failed to open ELF binary: %v", err)
-	}
-	defer f.Close()
-
-	candidates, err := resurgo.DetectFunctionsFromELF(f)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if len(candidates) == 0 {
-		t.Fatal("expected at least one function candidate, got none")
-	}
-
-	counts := make(map[resurgo.DetectionType]int)
-	for _, c := range candidates {
-		counts[c.DetectionType]++
-	}
-	t.Logf("total candidates: %d, by type: %v", len(candidates), counts)
-
-	if counts[resurgo.DetectionPrologueOnly] == 0 {
-		t.Error("expected at least one prologue-only candidate")
-	}
-}
-
-// TestDisasmDetector verifies that DisasmDetector, when run against a real ELF
-// binary, produces candidates with the expected detection types and that
-// functions both called and matching a prologue pattern are promoted to
-// DetectionPrologueCallSite with ConfidenceHigh.
-func TestDisasmDetector(t *testing.T) {
-	binPath := filepath.Join(t.TempDir(), "demo-app")
-	cmd := exec.Command("go", "build", "-o", binPath, "testdata/demo-app.go")
-	cmd.Env = append(os.Environ(), "CGO_ENABLED=0", "GOARCH=amd64")
-	if out, err := cmd.CombinedOutput(); err != nil {
-		t.Fatalf("failed to compile demo-app: %v\n%s", err, out)
-	}
-
-	f, err := elf.Open(binPath)
-	if err != nil {
-		t.Fatalf("failed to open ELF binary: %v", err)
-	}
-	defer f.Close()
-
-	candidates, err := resurgo.DisasmDetector(f)
-	if err != nil {
-		t.Fatalf("DisasmDetector: %v", err)
-	}
-
-	if len(candidates) == 0 {
-		t.Fatal("expected at least one candidate, got none")
-	}
-
-	counts := make(map[resurgo.DetectionType]int)
-	for _, c := range candidates {
-		counts[c.DetectionType]++
-	}
-	t.Logf("total candidates: %d, by type: %v", len(candidates), counts)
-
-	// Disasm must find functions via prologue pattern.
-	if counts[resurgo.DetectionPrologueOnly] == 0 && counts[resurgo.DetectionPrologueCallSite] == 0 {
-		t.Error("expected prologue-based candidates, got none")
-	}
-
-	// Functions confirmed by both prologue and call-site must be ConfidenceHigh
-	// and must carry at least one caller address.
-	for _, c := range candidates {
-		if c.DetectionType == resurgo.DetectionPrologueCallSite {
-			if c.Confidence != resurgo.ConfidenceHigh {
-				t.Errorf("0x%x: expected ConfidenceHigh for prologue-callsite, got %s", c.Address, c.Confidence)
-			}
-			if len(c.CalledFrom) == 0 && len(c.JumpedFrom) == 0 {
-				t.Errorf("0x%x: prologue-callsite candidate has no caller or jump source", c.Address)
-			}
-		}
-	}
-}
-
-func TestDetectFunctionsFromELF_InvalidELF(t *testing.T) {
-	r := bytes.NewReader([]byte{0x00, 0x01, 0x02, 0x03})
-	f, err := elf.NewFile(r)
-	if err == nil {
-		f.Close()
-		t.Fatal("expected elf.NewFile to fail on invalid data")
 	}
 }
